requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
fastapi-cache2==0.2.1
redis==4.6.0
aioredis==1.3.1
celery==5.3.6
shapely==2.0.2
requests==2.22.0
pymysql==1.1.0
cachetools==5.3.2
backoff==2.2.1
httpx==0.26.0
starlette==0.32.0.post1